Split FunctionBinder from FunctionResolver
